feat: [CVNG-39]: data source panel and etrics component

+++ src/modules/cv/components/DataSourceConfigPanel/DataSourceConfigPanel.tsx
@@ -6,15 +6,20 @@ import type { DSConfig } from '@wings-software/swagger-ts/definitions'
 import css from './DataSourceConfigPanel.module.scss'
 import type { ListPanelInterface } from '@wings-software/uikit/dist/components/Collapse/CollapseListPanel'
 import { useFormikContext } from 'formik'
+import i18n from './DataSourceConfigPanel.i18n'
 
-async function removeDataSourceConfig(accountId: string, idToDelete?: string): Promise<string | undefined> {
-  if (!idToDelete) {
-    return
-  }
+async function removeDataSourceConfig(
+  accountId: string,
+  dataSourceConnectorId: string,
+  identifier: string,
+  productName?: string
+): Promise<string | undefined> {
   const { error } = await CVNextGenCVConfigService.deleteConfigs({
     accountId,
-    group: 'XHR_DELETE_CONFIG_GROUP',
-    configsToDelete: [idToDelete]
+    dataSourceConnectorId,
+    identifier,
+    productName,
+    group: 'XHR_DELETE_CONFIG_GROUP'
   })
   return error ? error : undefined
 }
@@ -42,13 +47,24 @@ interface DataSourceConfigPanelProps extends ListPanelInterface {
 }
 
 const DataSourceConfigPanel: React.FC<DataSourceConfigPanelProps> = (props): JSX.Element => {
-  const { entityName, onRemove, children, transformToSavePayload, index, isOpen, onToggleOpen, validate } = props
+  const {
+    entityName,
+    onRemove,
+    children,
+    transformToSavePayload,
+    index,
+    isOpen,
+    onToggleOpen,
+    validate,
+    openNext
+  } = props
   const [panelHeaderMsg, setPanelHeaderMsg] = useState<{ intent: Intent; msg: string } | undefined>()
   const [hasSaved, setSaved] = useState<boolean>()
   const { values, setFieldTouched, setFieldError } = useFormikContext<{
     dsConfigs: DSConfig[]
   }>()
   const configData = values?.dsConfigs?.[index] || {}
+  const accountId = configData.accountId
 
   const headingComp = useMemo(
     () => (
@@ -72,7 +88,7 @@ const DataSourceConfigPanel: React.FC<DataSourceConfigPanelProps> = (props): JSX
       setFieldError(`dsConfigs[${index}].${field}`, panelErrors[field])
     })
     if (Object.keys(panelErrors).length) {
-      setPanelHeaderMsg({ intent: 'danger', msg: 'One or more fields are invalid.' })
+      setPanelHeaderMsg({ intent: 'danger', msg: i18n.invalidMessage })
       return false
     }
     return true
@@ -81,54 +97,69 @@ const DataSourceConfigPanel: React.FC<DataSourceConfigPanelProps> = (props): JSX
   const onToggleOpenCallback = useCallback(
     (hasOpened?: boolean) => {
       if (!hasOpened && validateCallback() && !hasSaved && panelHeaderMsg && panelHeaderMsg.intent === 'primary') {
-        setPanelHeaderMsg({ intent: 'none', msg: 'Unsaved Changes' })
+        setPanelHeaderMsg({ intent: 'none', msg: i18n.usavedChangesMessage })
       }
-      onToggleOpen(hasOpened)
+      onToggleOpen?.(hasOpened)
     },
     [hasSaved, onToggleOpen, panelHeaderMsg, validateCallback]
   )
 
   const onRemoveCallback = useCallback(() => {
-    removeDataSourceConfig(configData.accountId || '').then(error => {
-      if (!error) {
-        setPanelHeaderMsg(undefined)
-        onRemove(index)
-      } else {
-        setPanelHeaderMsg({ intent: 'danger', msg: error })
-      }
-    })
-  }, [onRemove, configData, index])
+    if (!configData.identifier && !panelHeaderMsg) {
+      onRemove(index)
+      return
+    }
+
+    if (!window.confirm('Are you sure you want to delete the item?')) {
+      return
+    }
+    if (configData.identifier) {
+      removeDataSourceConfig(
+        accountId || '',
+        configData.connectorId || '',
+        configData.identifier || '',
+        configData.productName
+      ).then(error => {
+        if (!error) {
+          setPanelHeaderMsg(undefined)
+          onRemove(index)
+        } else {
+          setPanelHeaderMsg({ intent: 'danger', msg: error })
+        }
+      })
+    } else {
+      onRemove(index)
+    }
+  }, [onRemove, accountId, index, configData, panelHeaderMsg])
 
   const collapseHeaderProps: CollapseListPanelProps['collapseHeaderProps'] = useMemo(
     () => ({
       heading: headingComp,
       isRemovable: true,
       onRemove: onRemoveCallback
-      // onToggleOpen: () => null,
-      // isOpen: isOpen || false
     }),
-    [headingComp, isOpen, onRemoveCallback]
+    [headingComp, onRemoveCallback]
   )
 
-  const onNextCallback = useCallback(() => {
+  const onNextCallback = useCallback(async () => {
+    // if (!validateCallback()) {
+    //   return
+    // }
     const config: DSConfig = transformToSavePayload?.(configData) || configData
-    saveDSConfig(config, config.accountId || '').then(error => {
-      if (!validateCallback()) {
-        return
-      }
-      if (error) {
-        setPanelHeaderMsg({ msg: error, intent: 'danger' })
-      } else {
-        setPanelHeaderMsg({ msg: 'Success', intent: 'success' })
-      }
-    })
-  }, [configData, transformToSavePayload, validateCallback])
+    const error = await saveDSConfig(config, config.accountId || '')
+    if (error) {
+      setPanelHeaderMsg({ msg: error, intent: 'danger' })
+    } else {
+      setPanelHeaderMsg({ msg: i18n.successMessage, intent: 'success' })
+    }
+    openNext?.()
+  }, [configData, transformToSavePayload, validateCallback, openNext])
 
   useEffect(() => {
     if (!isOpen) {
       return
     }
-    setPanelHeaderMsg({ intent: 'primary', msg: 'Editing' })
+    setPanelHeaderMsg({ intent: 'primary', msg: i18n.editingMessage })
     if (hasSaved) {
       setSaved(false)
     }
@@ -140,7 +171,7 @@ const DataSourceConfigPanel: React.FC<DataSourceConfigPanelProps> = (props): JSX
       onToggleOpen={onToggleOpenCallback}
       isOpen={isOpen || false}
       className={css.main}
-      nextButtonText="Save"
+      nextButtonText={i18n.saveButtonText}
       openNext={onNextCallback}
     >
       {children}
